make auth.py for login and logout functionality
